requirements.txt
asgiref==3.2.10
certifi==2020.6.20
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1
djangorestframework==3.11.0
drf-yasg==1.20.0
markdown==3.3.4
django-filter==2.4.0
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.4
psycopg2==2.8.5
sqlparse==0.3.1
Unipath==1.1
pytest==5.4.0
faker==9.6.0
gunicorn==20.1.0
dj-static==0.0.6